+++ HISTORY.md
@@ -2,6 +2,7 @@ History
 =======
 
 ## Current
+* Change `host` short flag to `-H` (don't conflict with help `-h`).
 * Add Handlebars template support.
 
 ## v0.0.4